Generate beta hats from standardized or non-standardized direct SNP effects and LD

Generate beta hats from standardized or non-standardized direct SNP effects and LD

Usage

gen_bhat_from_b(
  b_joint_std,
  b_joint,
  N,
  trait_corr = NULL,
  R_LD = NULL,
  af = NULL,
  est_s = FALSE,
  L_mat_joint_std = NULL,
  theta_joint_std = NULL
)

Arguments

b_joint_std

Matrix of standardized joint (causal) effects (dimension variants by traits)

b_joint

Matrix of non-standardized joint (causal) effects (dimension variants by traits). Supply only one of b_joint or b_joint_std.

N

Sample size, scalar, vector, or matrix. See ?sim_mv for more details.

trait_corr

Matrix of population trait correlation (traits by traits)

R_LD

LD pattern (optional). See ?sim_mv for more details.

af

Allele frequencies (optional, allowed only if R_LD is missing). See ?sim_mv for more details.

Details

This function can be used to generate new GWAS results with the same effect sizes by passing in the beta_joint table from a data set simulated using `sim_mv`. If the original data are generated with af missing and no LD then the beta_joint table contains standardized effects. Otherwise it contains non-standardized effects. Use the appropriate argument, either b_joint_std or b_joint.

Examples

# Use gen_bhat_from_b to generate new GWAS results with the same effect sizes.
N <- matrix(1000, nrow = 2, ncol =2)
G <- matrix(0, nrow = 2, ncol = 2)
R_E <- matrix(c(1, 0.8, 0.8, 1), nrow = 2, ncol = 2)
# original data
dat <- sim_mv(N = N, J = 20000, h2 = c(0.4, 0.3), pi = 1000/20000,
               G = G, R_E = R_E)
#> SNP effects provided for 20000 SNPs and 2 traits.
# data for second GWAS
# Since we didn't supply af or an LD pattern in the original GWAS,
# we have standardized effects.
dat_new <- gen_bhat_from_b(b_joint_std = dat$beta_joint, N = 40000,
                           trait_corr = dat$trait_corr)
#> SNP effects provided for 20000 SNPs and 2 traits.
#> Warning: trait_corr disregarded as there is no sample overlap.