Sample individual level data with joint effects matching a sim_mv object

Sample individual level data with joint effects matching a sim_mv object

Usage

resample_inddata(
  sim_dat,
  N,
  R_LD = NULL,
  af = NULL,
  sim_func = gen_genos_mvn,
  calc_sumstats = FALSE
)

Arguments

sim_dat

An object of class sim_mv (produced by sim_mv or gen_bhat_from_b).

N

Sample size, scalar, vector, or special sample size format data frame, see details.

R_LD

LD pattern (optional). See ?sim_mv for more details.

af

Allele frequencies. This can be a scalar, vector or a function. For this function, af must always be supplied.

V_E

Vector with length equal to the number of traits giving the environmental variance of each trait.

R_E

Environmental correlation matrix, (traits by traits). If missing, R_E is assumed to be the identity.

Details

This function can be used to generate individual level GWAS data by passing in the beta_joint table from a data set simulated using `sim_mv`. If the original data are generated with af missing and no LD then the beta_joint table contains standardized effects. Otherwise it contains non-standardized effects. Use the appropriate argument, either b_joint_std or b_joint.

Examples

# Use gen_gwas_from_b to generate individual level data with given effect size.
Ndf <- data.frame(trait_1 = 1, trait_2 = 1, N = 10000)
G <- matrix(0, nrow = 2, ncol = 2)
R_E <- matrix(c(1, 0.8, 0.8, 1), nrow = 2, ncol = 2)
# original data
dat <- sim_mv(N = Ndf, J = 2000, h2 = c(0.4, 0.3), pi = 100/2000,
               G = G, R_E = R_E, af = function(n){rbeta(n, 1, 5)})
#> SNP effects provided for 2000 SNPs and 2 traits.
# Now generate GWAS data
gw_dat <- resample_inddata(dat, N = Ndf, calc_sumstats = TRUE)
#> SNP effects provided for 2000 SNPs and 2 traits.
#> Error in resample_inddata(dat, N = Ndf, calc_sumstats = TRUE): Must supply allele frequencies (af).